+++ BetterZeeLog/Patches/UIGroupHandlerPatch.cs
@@ -5,43 +5,30 @@ using System.Reflection.Emit;
 
 using HarmonyLib;
 
-using static PluginConfig;
-
 [HarmonyPatch(typeof(UIGroupHandler))]
 static class UIGroupHandlerPatch {
   [HarmonyTranspiler]
   [HarmonyPatch(nameof(UIGroupHandler.Update))]
-  static IEnumerable<CodeInstruction> UpdateTranspiler(
+  static IEnumerable<CodeInstruction> Update1Transpiler(
       IEnumerable<CodeInstruction> instructions, ILGenerator generator) {
     return new CodeMatcher(instructions, generator)
         .Start()
         .MatchStartForward(
             new CodeMatch(OpCodes.Call, AccessTools.Method(typeof(ZLog), nameof(ZLog.Log))))
         .ThrowIfNotMatch("Could not patch UIGroupHandler.Update()! (log-right-stick)")
-        .Advance(offset: 1)
-        .CreateLabel(out Label activateRightStickLabel)
+        .SetInstruction(
+            new CodeInstruction(
+                OpCodes.Call, AccessTools.Method(typeof(UIGroupHandlerPatch), nameof(LogDelegate))))
         .MatchStartForward(
             new CodeMatch(OpCodes.Call, AccessTools.Method(typeof(ZLog), nameof(ZLog.Log))))
         .ThrowIfNotMatch("Could not patch UIGroupHandler.Update()! (log-default)")
-        .Advance(offset: 1)
-        .CreateLabel(out Label activeDefaultLabel)
-        .Start()
-        .MatchStartForward(
-            new CodeMatch(OpCodes.Ldstr, "Activating right stick element "))
-        .ThrowIfNotMatch($"Could not patch UIGroupHandler.Update()! (ldstr-right-stick)")
-        .InsertAndAdvance(
-            Transpilers.EmitDelegate(ShouldLogDelegate),
-            new CodeInstruction(OpCodes.Brfalse, activateRightStickLabel))
-        .MatchStartForward(
-            new CodeMatch(OpCodes.Ldstr, "Activating default element "))
-        .ThrowIfNotMatch($"Could not patch UIGroupHandler.Update()! (ldstr-default)")
-        .InsertAndAdvance(
-            Transpilers.EmitDelegate(ShouldLogDelegate),
-            new CodeInstruction(OpCodes.Brfalse, activeDefaultLabel))
+        .SetInstruction(
+            new CodeInstruction(
+                OpCodes.Call, AccessTools.Method(typeof(UIGroupHandlerPatch), nameof(LogDelegate))))
         .InstructionEnumeration();
   }
 
-  static bool ShouldLogDelegate() {
-    return !IsModEnabled.Value;
+  static void LogDelegate(object obj) {
+    // ...
   }
 }